requirements.txt
fastapi==0.95.2
starlette==0.27.0
pydantic==1.10.12
typing-extensions==4.5.0
uvicorn==0.22.0
pytest==7.4.3
httpx==0.24.1
anyio==3.6.2
jinja2==3.1.2
python-multipart==0.0.6
requests==2.31.0
python-dotenv==1.0.0
twilio==8.10.0

# 네트워크 제어 관련 의존성
# RADIUS 클라이언트 (선택사항)
# pyrad==2.4

# 네트워크 모니터링 (선택사항)
# psutil==5.9.5

# 암호화 및 보안
cryptography==41.0.7
